third_party/blink/web_tests/fast/dom/Window/window-open-no-multiple-windows-from-iframe-expected.txt

CONSOLE ERROR: line 4: Not allowed to navigate top frame to data URL: data:text/html;charset=utf-8,<html><body>PASS</body></html>
PASS
